==> 01_features.ipynb <==
import pandas as pd
file = pd.read_csv('events_up_to_01062018.csv')

file['timestamp'] = pd.to_datetime(file['timestamp'])

## Saco features

#cantidad de compras, visitas, productos vistos
def cantidades (x):
    compras = (x['event']=='conversion').sum()
    visitas = (x['event']=='visited site').sum()
    cant_prod_vistos = (x['event']=='viewed product').sum()
    return pd.Series([compras,visitas,cant_prod_vistos])
cantidades = file.groupby('person').apply(cantidades)
cantidades.columns=['compras','visitas','cant_prod_vistos']

#el tipo de dispositivo que mas uso para entrar
def dispos (x):
    dispos = x['device_type'].value_counts().index
    if dispos.size>0:
        dispositivo = dispos[0]
    else:
        dispositivo = 'Ninguno'
    return pd.Series([dispositivo])
dispos = file.groupby('person').apply(dispos)
dispos.columns=['dispos']

#el producto que mas vio
def prodvisto (x):
    prods = x.loc[x.event=='viewed product','model'].value_counts().index
    if prods.size>0:
        prod_visto = prods[0]
    else:
        prod_visto = 'Ninguno'
    return pd.Series([prod_visto])
prodvisto = file.groupby('person').apply(prodvisto)
prodvisto.columns = ['prod_visto']

#el evento que realizo mas veces
def mayorevento (x):
    mayor_evento = x['event'].value_counts().index[0]
    return pd.Series([mayor_evento])
mayorevento = file.groupby('person').apply(mayorevento)
mayorevento.columns = ['mayor_evento']

features = mayorevento.join(prodvisto).join(dispos).join(cantidades)

#features temporales
def temporal (x):
    s = x['timestamp'].size
    primer_visita = x['timestamp'].min()
    ultima_visita = x['timestamp'].max()
    cant_eventos = s
    return pd.Series([primer_visita,ultima_visita,cant_eventos])
temporal = file.groupby('person').apply(temporal)

#channel por el que mas entró
def channel (x):
    channels = x['channel'].value_counts()
    if channels.size > 0:
        return channels.index[0]
    else:
        return 'Ninguno'
channel = file.groupby('person').apply(channel)
channel.columns = ['channel']

temporal.columns = ['primer_visita','ultima_visita','cant_eventos']
processed = features.join(temporal)

processed.head()

## Un par de features mas a partir de las que ya saque:

processed['permanencia'] = (processed['ultima_visita']-processed['primer_visita']).dt.days + 1

processed['hace_cuanto_visito'] = (file['timestamp'].max()-processed['ultima_visita']).dt.days

processed['eventos_por_dia'] = processed['cant_eventos']/processed['permanencia']

processed['channel'] = channel

processed['visitas_por_dia'] = processed['visitas']/processed['permanencia']

## Lo guardo como csv para usarlo en la prediccion

processed.drop(['primer_visita','ultima_visita'],axis=1).to_csv('features_01.csv')
