==> 08_features_nuevas.ipynb <==
import pandas as pd

file = pd.read_csv('events_up_to_01062018.csv',low_memory=False)

file.info()

def isnew (x):
    c = x['new_vs_returning'].value_counts().index
    return ('New' in c)
new = file.groupby('person').apply(isnew)

new.value_counts()

def region (x):
    c = x['region'].value_counts().index
    if len(c)>0:
        return c[0]
    else:
        return 'Unknown'
reg = file.groupby('person').apply(region)

reg.head()

def viocelulares(x):
    return (x.loc[x.event=='viewed product']['event'].size > 0)
viocelulares = file.groupby('person').apply(viocelulares)

viocelulares.value_counts()

features = pd.DataFrame(viocelulares).join(pd.DataFrame(reg),lsuffix='v').join(pd.DataFrame(new))
features.columns = ['vio_celulares','region','is_new']

features.head()

features_01 = pd.read_csv('features_01.csv')

features_03 = features_01.set_index('person').join(features,how='left')

features_03.head()

features_03.to_csv('features_03.csv')
