==> 09_xgboost.ipynb <==
import numpy as np
import scipy.stats as stats

import xgboost as xgb

import pandas as pd
from sklearn.model_selection import train_test_split
import category_encoders

from sklearn.model_selection import RandomizedSearchCV

predict = pd.read_csv('trocafone_kaggle_test.csv')

a_predecir = pd.read_csv('labels_training_set.csv')

processed = pd.read_csv('features_03.csv')

processed = processed.set_index('person')
processed.head()

predict_features = processed.loc[processed.index.isin(predict['person'])].drop(['primer_visita','ultima_visita'],axis=1)

training_features = processed.loc[processed.index.isin(a_predecir['person'])]
training_features['label'] = a_predecir.set_index('person')['label']


training_labels = training_features['label']
training_features = training_features.drop(['primer_visita','ultima_visita','label'],axis=1)

training_features.head()

rfilter = (training_features['prod_visto']!='Ninguno') & (training_features['dispos']!='Ninguno') & (training_features['channel']!='Ninguno') & (training_features['region']!='Unknown')

filtered_training = training_features.loc[rfilter]
filtered_labels = training_labels.loc[rfilter]

## Mean encoding

tg = category_encoders.TargetEncoder(smoothing=True)
encoded_training = tg.fit_transform(np.array(training_features),np.array(training_labels))
encoded_predict = tg.transform(np.array(predict_features))

encoded_training.head()

params={
    'n_estimators':stats.randint(150,500),
    'learning_rate':stats.uniform(0.01,0.3),
    'subsample':stats.uniform(0.3,0.7),
    'min_child_weight':[1,5,10],
    'max_depth':[3,10,6],
    'gamma':stats.randint(0,10),
    'colsample_bytree':stats.uniform(0.4,0.6)
}
grid = RandomizedSearchCV(xgb.XGBClassifier(n_jobs=-1),
                          param_distributions=params,
                          scoring='roc_auc',
                          cv=2,
                          verbose=1,
                          n_iter=100)

xtr,xte,ytr,yte=train_test_split(encoded_training,training_labels)

grid.fit(xtr,ytr)

grid.score(X=xte,y=yte)

grid.best_params_

## Prediccion

predictor = grid.best_estimator_

predictor.feature_importances_

predictor.fit(encoded_training,training_labels)

prediction_submit = predictor.predict_proba(encoded_predict)[:,1]

predict_features['label'] = prediction_submit

subm = predict.merge(predict_features['label'].reset_index(),how='left',on='person')

subm = subm.set_index('person')

subm.to_csv('submit.csv')
