==> 11_features.ipynb <==
import pandas as pd
file = pd.read_csv('events_up_to_01062018.csv',low_memory=False)

file.info()

features_03 = pd.read_csv('features_03.csv')
features_03.head()

features_03.info()

f1 = features_03.groupby('prod_visto').agg({'person':'count'})

features_04 = features_03.merge(f1,how='left',on='prod_visto')

def scrnres (x):
    c = x['screen_resolution'].value_counts().index
    if len(c)>0:
        return c[0]
    else:
        return 'Ninguno'
scrnres = file.groupby('person').apply(scrnres)
scrnres = pd.DataFrame(scrnres)
scrnres.columns = ['scrnres']

def opsis (x):
    c = x['operating_system_version'].value_counts().index
    if len(c)>0:
        return c[0]
    else:
        return 'Ninguno'
opsis = file.groupby('person').apply(opsis)
opsis = pd.DataFrame(opsis)
opsis.columns = ['opsis']

opsis.head()

features_04 = features_04.set_index(features_04.columns[0]).join(opsis).join(scrnres)

features_04 = features_04.reset_index()

features_04.head()

a = ['mayor_evento', 'dispos','channel','region']
to_merge = []
for c in a:
    b = features_04.groupby(c).agg({features_04.columns[0]:'count'})
    features_04 = features_04.merge(b,how='left',on=c)

features_04.head()

def channel_inicial (x):
    vsc = x.loc[x.event=='visited site']['channel']
    if vsc.size>0:
        return vsc.values[0]
    else:
        return 'Unknown'
ci = file.groupby('person').apply(channel_inicial)

ci.value_counts()

features_04 = features_04.set_index(features_04.columns[0]).join(pd.DataFrame(ci)).reset_index()

features_04.head()

file.event.value_counts()

def notificaciones (x):
    return x.loc[x.event=='lead'].index.size
notif = file.groupby('person').apply(notificaciones)
notif = pd.DataFrame(notif)
notif.columns = ['lead']

features_04 = features_04.set_index(features_04.columns[0]).join(notif).reset_index()

features_04.head()

a = ['opsis','scrnres']
to_merge = []
for c in a:
    b = features_04.groupby(c).agg({features_04.columns[0]:'count'})
    features_04 = features_04.merge(b,how='left',on=c)

features_04.head()

file.event.unique()

tipos_de_eventos = list(file['event'].unique())
tipos_de_eventos.remove('viewed product') 
tipos_de_eventos.remove('staticpage') #la feature de si vio celulares o no ya se hizo y staticpage no tiene mucho sentido
def hizo_eventos(x):
    feats = []
    for event in tipos_de_eventos:
        f = event in x['event'].unique()
        feats.append(f)
    return pd.Series(feats)
hizoeventos = file.groupby('person').apply(hizo_eventos)
hizoeventos.columns = ['busco_productos', 'clickeo_ad','hizo_checkout', 'entro_desde_motor', 'hizo_conversion','filtro_busquedas', 'busco_marca', 'visito_sitio', 'hizo_lead']

hizoeventos.head()

features_05 = features_04.set_index(features_04.columns[0]).join(hizoeventos,how='left').reset_index()

features_05.head()

features_05.columns

features_05.columns = ['person',         'mayor_evento',           'prod_visto',
                     'dispos',              'compras',              'visitas',
           'cant_prod_vistos',        'primer_visita',        'ultima_visita',
               'cant_eventos',          'permanencia',   'hace_cuanto_visito',
            'eventos_por_dia',              'channel',  'visitas_por_dia',      'vio_celulares',
                     'region',               'is_new',             'personas_mismo_prod_visto',
                      'opsis',              'scrnres',           'personas_mismo_mayor_evento',
               'personas_mismo_disp',       'personas_mismo_region',     'personas_mismo_channel',
                            'primer_channel',                 'lead',   'personas_mismo_opsis',
       'personas_mismo_scrnres',      'busco_productos',           'clickeo_ad',
              'hizo_checkout',    'entro_desde_motor',      'hizo_conversion',
           'filtro_busquedas',          'busco_marca',         'visito_sitio',
                  'hizo_lead']

features_05.set_index(features_05.columns[0]).to_csv('features_04.csv')

features_05.info()
