==> 16_LGBM.ipynb <==
import numpy as np
import scipy.stats as stats

import lightgbm as lgbm

import pandas as pd
from sklearn.model_selection import train_test_split
import category_encoders

from sklearn.model_selection import RandomizedSearchCV

predict = pd.read_csv('trocafone_kaggle_test.csv')

a_predecir = pd.read_csv('labels_training_set.csv')

processed = pd.read_csv('features_04.csv')

cfeats = ['mayor_evento', 'prod_visto', 'dispos',  'channel', 'region', 'opsis', 'scrnres','cuantos_leads']
for c in cfeats:
    processed[c] = processed[c].astype('category')

processed = processed.set_index(processed.columns[0])
processed.head()

predict_features = processed.loc[processed.index.isin(predict['person'])].drop(['primer_visita','ultima_visita'],axis=1)

predict_features.head()

training_features = processed.loc[processed.index.isin(a_predecir['person'])]
training_features['label'] = a_predecir.set_index('person')['label']


training_labels = training_features['label']
training_features = training_features.drop(['primer_visita','ultima_visita','label'],axis=1)

training_features.head()

params = {
    'num_leaves':stats.randint(90,5000),
    'min_data_in_leaf':stats.randint(100,1000),
    'max_bin':stats.randint(70,300),
    'learning_rate':stats.uniform(0.01,1-0.01),
    'n_estimators':stats.randint(150,500)
}
grid = RandomizedSearchCV(lgbm.LGBMClassifier(boosting_type='dart'),
                          param_distributions=params,
                          scoring='roc_auc',
                          cv=2,
                          verbose=1,
                          n_iter=150)

xtr,xte,ytr,yte=train_test_split(training_features,training_labels)

grid.fit(xtr,ytr)

grid.score(X=xte,y=yte)

grid.best_params_

## Prediccion

predictor = grid.best_estimator_

predictor.feature_importances_

predictor.fit(training_features,training_labels)

prediction_submit = predictor.predict_proba(predict_features)[:,1]

predict_features['label']=prediction_submit

submit = predict.set_index('person')

submit['label'] = predict_features['label']

submit.to_csv('submit.csv')
