==> 17_xgboost-catboost-mean.ipynb <==
## Score en kaggle: 0.85481

import catboost as cb
import xgboost as xgb
import pandas as pd

import scipy.stats as stats
import numpy as np

from sklearn.model_selection import train_test_split
import category_encoders

from sklearn.model_selection import RandomizedSearchCV

kaggle = pd.read_csv('trocafone_kaggle_test.csv')

labels = pd.read_csv('labels_training_set.csv')

processed = pd.read_csv('features_04.csv')

processed = processed.set_index(processed.columns[0])
processed = processed.drop(['primer_visita','ultima_visita'],axis=1)
processed.head()

# XGBoost

training_features = labels.set_index('person').join(processed,how='left')
predict_features = kaggle.set_index('person').join(processed,how='left')

training_labels = training_features['label']
training_features = training_features.drop(['label'],axis=1)

## Mean encoding

tg = category_encoders.TargetEncoder(smoothing=True)
encoded_training = tg.fit_transform(np.array(training_features),np.array(training_labels))
encoded_predict = tg.transform(np.array(predict_features))

encoded_training.head()

## Prediccion (pre-entrenado)

predictor = xgb.XGBClassifier()

predictor.load_model('preentrenados/85193xgboost')

prediction_submit = predictor.predict_proba(encoded_predict)[:,1]

predict_features['label']=prediction_submit

submit_xgb = kaggle.set_index('person')

submit_xgb['label'] = predict_features['label']

# CatBoost

catBoost = cb.CatBoostClassifier()
catBoost.load_model('preentrenados/85317catboost')

pred = catBoost.predict_proba(predict_features)[:,1]
submit_cb = kaggle.set_index('person')
submit_cb['label'] = pred

submit_cb.head()

submit_xgb.head()

submit = 0.6*submit_cb+0.4*submit_xgb

submit.to_csv('submit.csv')
