==> 18_features.ipynb <==
import pandas as pd
import numpy as np

file = pd.read_csv('events_up_to_01062018.csv',low_memory=False)

file.info()

gbp = file.groupby('person')

gbp

features_04 = pd.read_csv('features_04.csv')

## Primer evento

first_event = gbp.agg({'event':'first'})
first_event.columns = ['primer_evento']

## Modelos de celular distintos que vio

def modelosdistintos (x):
    return x.loc[x.event=='viewed product']['model'].nunique()
mod = gbp.apply(modelosdistintos)

## Cuantas veces hizo cada evento

events = ['searched products', 'ad campaign hit','checkout', 'search engine hit','generic listing', 'brand listing']
def cadaevento (x):
    l = []
    for e in events:
        l.append (x.loc[x.event==e,'event'].size)
    return pd.Series(l)
cadaevento = gbp.apply(cadaevento)

cadaevento.columns = ['busquedas','clicks_en_ads','checkouts','entradas_desde_motor','listar_productos','listar_marca']

## Tamaño de pantalla numerico (ancho x alto)

tampant = features_04.scrnres.str.split('x').apply(lambda x: (int(x[0])*int(x[1])) if x[0]!='Ninguno' else 0  )

## Entro desde 'direct'

direct = features_04['primer_channel'] == 'Direct'

## Entro desde 'Computer'

computer = features_04['dispos'] == 'Computer'

features_05 = features_04.set_index('person').join(first_event,how='left')

features_05['modelos_distintos'] = mod
features_05 = features_05.join(cadaevento,how='left')
features_05['tamanio_pantalla'] = tampant
features_05['entro_direct'] = direct
features_05['uso_computer'] = computer

features_05.head()

features_05.to_csv('features_05.csv')
