==> 19_catboost.ipynb <==
import catboost as cb
import pandas as pd
from sklearn.model_selection import GridSearchCV
from sklearn.model_selection import train_test_split
import numpy as np
from sklearn.metrics import roc_auc_score

labels = pd.read_csv('labels_training_set.csv')

processed = pd.read_csv('features_05.csv')

kaggle = pd.read_csv('trocafone_kaggle_test.csv')

processed.head()

features = processed.set_index(processed.columns[0]).drop(['primer_visita','ultima_visita'],axis=1)

dataset = labels.set_index('person').join(features,how='left')

dataset.head()

for x in dataset.columns:
    if dataset[x].dtype == 'object':
        dataset[x]=dataset[x].astype('category')

training_features = dataset.drop(['label'],axis=1)
training_labels = dataset['label']

xtrain,xtest,ytrain,ytest = train_test_split(training_features,training_labels)

catBoost = cb.CatBoostClassifier(iterations=200,
                                 depth=8,
                                 l2_leaf_reg=50,
                                 random_strength=0.1,
                                 bagging_temperature=0,
                                 border_count=128)

cat = []
for index,v in enumerate(xtrain.columns):
    if xtrain[v].dtype.name == 'category':
        cat.append(index)


catBoost.fit(xtrain,ytrain,verbose=0,cat_features=cat)

roc_auc_score(ytest,catBoost.predict_proba(xtest)[:,1])

catBoost.fit(training_features,training_labels,verbose=0,cat_features=cat)

prediction_features = kaggle.set_index('person').join(features,how='left')

pred = catBoost.predict_proba(prediction_features)[:,1]
submit = kaggle.set_index('person')
submit['label'] = pred

submit.head()

submit.to_csv('submit.csv')
