==> 20_Stacked.ipynb <==
import pandas as pd
from random import sample

import numpy as np
import category_encoders as ce

from sklearn.model_selection import train_test_split
from sklearn.metrics import roc_auc_score

from sklearn.linear_model import LogisticRegression 
from sklearn.ensemble import RandomForestClassifier

import xgboost as xgb
import lightgbm as lgb
import catboost as cb
from sklearn.neighbors import KNeighborsClassifier

from sklearn.model_selection import KFold

## Train y predict set

file = pd.read_csv('features_05.csv')
file = file.drop(['primer_visita','ultima_visita'],axis=1)
labels = pd.read_csv('labels_training_set.csv')
kaggle = pd.read_csv('trocafone_kaggle_test.csv')

training = labels.set_index('person').join(file.set_index('person'),how='left')

training_set = training.drop('label',axis=1)
training_labels = training['label']

prediction_set = kaggle.set_index('person').join(file.set_index('person'),how='left')

cat = []
catnames = []
for index,v in enumerate(training_set.columns):
    if training_set[v].dtype.name == 'object':
        cat.append(index)
        catnames.append(v)

for i in catnames:
    training_set[i] = training_set[i].astype('category')
    prediction_set[i] = prediction_set[i].astype('category')

## Funciones

def agregarFeatureTrain (dataset,labels,clasif_list,n):
    to_return = []
    new_dataset = dataset.copy(deep=True)
    kf = KFold(n_splits=n)
    new_feature = []
    for clasificador in clasif_list:
        for a,b in kf.split(dataset):
            clasificador.fit(dataset.iloc[a],labels.iloc[a])
            new_labels = clasificador.predict_proba(dataset.iloc[b])[:,1]
            new_feature.extend(new_labels)
        newindex = len(new_dataset.columns)+1
        new_dataset[newindex] = new_feature
        to_return.append(newindex)
    return new_dataset.loc[:,to_return]

def agregarFeatureTrainCB (dataset,labels,clasif_list,n):
    to_return = []
    new_dataset = dataset.copy(deep=True)
    kf = KFold(n_splits=n)
    new_feature = []
    for clasificador in clasif_list:
        for a,b in kf.split(dataset):
            clasificador.fit(dataset.iloc[a],labels.iloc[a],cat_features=cat)
            new_labels = clasificador.predict_proba(dataset.iloc[b])[:,1]
            new_feature.extend(new_labels)
        newindex = len(new_dataset.columns)+1
        new_dataset[newindex] = new_feature
        to_return.append(newindex)
    return new_dataset.loc[:,to_return]

def agregarFeaturePredict(dataset,labels,predict,clasif_list):
    to_return = []
    new_predict = predict.copy(deep=True)
    for classifier in clasif_list:
        classifier.fit(dataset,labels)
        prediction = classifier.predict_proba(predict)[:,1]
        newindex = len(predict.columns)+1
        new_predict[newindex] = prediction
        to_return.append(newindex)
    return new_predict.loc[:,to_return]

def agregarFeaturePredictCB(dataset,labels,predict,clasif_list):
    to_return = []
    new_predict = predict.copy(deep=True)
    for classifier in clasif_list:
        classifier.fit(dataset,labels,cat_features=cat)
        prediction = classifier.predict_proba(predict)[:,1]
        newindex = len(predict.columns)+1
        new_predict[newindex] = prediction
        to_return.append(newindex)
    return new_predict.loc[:,to_return]

## Clasificadores


lgbmparams = {   }
cb_classifier = cb.CatBoostClassifier(iterations=200,
                                 depth=8,
                                 l2_leaf_reg=50,
                                 random_strength=0.1,
                                 bagging_temperature=0,
                                 border_count=128)
xgb_classifier = xgb.XGBClassifier(colsample_bytree = 0.606653667095343,
             gamma = 7,
             learning_rate = 0.016238752458245277,
             max_depth = 10,
             min_child_weight = 5,
             n_estimators = 346,
             subsample = 0.7847190225361189)
lgbm_classifier = lgb.LGBMClassifier(learning_rate = 0.06748371705664348,
                 max_bin = 205,
                 min_data_in_leaf = 481,
                 n_estimators = 187,
                 num_leaves = 956)

## Encoding (para xgboost)

enc = ce.TargetEncoder()
encoded = enc.fit_transform(np.array(training_set),np.array(training_labels))
encoded_p = enc.transform(np.array(prediction_set))

## Obtener features

train_set1 = agregarFeatureTrain(encoded,training_labels,[xgb_classifier],4)
predict_set1 = agregarFeaturePredict(encoded,training_labels,encoded_p,[xgb_classifier])

train_set2 = agregarFeatureTrain(training_set,training_labels,[lgbm_classifier],4)
predict_set2 = agregarFeaturePredict(training_set,training_labels,prediction_set,[lgbm_classifier])

train_set3 = agregarFeatureTrainCB(training_set,training_labels,[cb_classifier],4)
predict_set3 = agregarFeaturePredictCB(training_set,training_labels,prediction_set,[cb_classifier])

train_set2.columns = ['lgbm']
train_set3.columns = ['catboost']

predict_set2.columns = ['lgbm']
predict_set3.columns = ['catboost']

## Armar stack (training)

stacked_training = train_set2.join(train_set3)
stacked_training['xgboost'] = list(train_set1[46])

stacked_training.head()

## Armar stack (prediction)

stacked_predict = predict_set2.join(predict_set3)
stacked_predict['xgboost'] = list(predict_set1[46])

stacked_predict.head()

## Prediccion final (Score: 0.85527)

##### Labels de xgboost, catboost y lightgbm con Logistic Regression como modelo final

log = LogisticRegression()

stacked_training.to_csv('preentrenados/training_logrec')

stacked_predict.to_csv('preentrenados/predict_logrec')

xtrain,xtest,ytrain,ytest = train_test_split(stacked_training,labels['label'])

log.fit(xtrain,ytrain)
roc_auc_score(ytest,log.predict_proba(xtest)[:,1])

log.fit(stacked_training,labels['label'])

final_pred = log.predict_proba(stacked_predict)[:,1]

submit = kaggle.copy(deep=True)
submit['label'] = final_pred

submit.set_index('person').to_csv('submit.csv')

## Usando algunas features del dataset 

fts = ['busco_productos', 'clickeo_ad', 'hizo_checkout', 'entro_desde_motor',
       'hizo_conversion', 'filtro_busquedas', 'busco_marca', 'visito_sitio',
       'hizo_lead', 'modelos_distintos', 'busquedas',
       'clicks_en_ads', 'checkouts', 'entradas_desde_motor',
       'listar_productos', 'listar_marca']

train_f = stacked_training.join(training_set.loc[:,fts])
pred_f = stacked_predict.join(prediction_set.loc[:,fts])

xtrain,xtest,ytrain,ytest = train_test_split(train_f,labels['label'])

log.fit(xtrain,ytrain)
roc_auc_score(ytest,log.predict_proba(xtest)[:,1])

log.fit(train_f,labels['label'])

n_prediction = log.predict_proba(pred_f)[:,1]

submit = kaggle.set_index('person').copy(deep=True)
submit['label'] = n_prediction

submit.to_csv('submit.csv')

## Xgboost con todas las features

forxgboost = stacked_training.reset_index().join(encoded).set_index('person')

xtrain,xtest,ytrain,ytest = train_test_split(forxgboost,labels['label'])

xgb_classifier.fit(xtrain,ytrain)
roc_auc_score(ytest,xgb_classifier.predict_proba(xtest)[:,1])

## Agrego KNN y RandomForests

##### KNN

from sklearn.preprocessing import normalize

normalized = pd.DataFrame(normalize(encoded,axis=0))
normalized_p = pd.DataFrame(normalize(encoded_p,axis=0))

knn = KNeighborsClassifier(leaf_size=15,p=3)

##### RandomForest

train_set_knn = agregarFeatureTrain(normalized,training_labels,[knn],4)
predict_set_knn = agregarFeaturePredict(normalized,training_labels,normalized_p,[knn])

rf = RandomForestClassifier(n_estimators=100)

train_set_rf = agregarFeatureTrain(encoded,training_labels,[rf],4)
predict_set_rf = agregarFeaturePredict(encoded,training_labels,encoded_p,[rf])

#### Armo set

stacked_features = train_set2.join(train_set3,how='left')
stacked_features['knn'] = list(train_set_knn[46])
stacked_features['rf'] = list(train_set_rf[46])
stacked_features['xgboost'] = list(train_set1[46])

stacked_features_p = predict_set2.join(predict_set3,how='left')
stacked_features_p['knn'] = list(predict_set_knn[46])
stacked_features_p['rf'] = list(predict_set_rf[46])
stacked_features_p['xgboost'] = list(predict_set1[46])

## Prediccion

log = LogisticRegression()

xtrain,xtest,ytrain,ytest = train_test_split(stacked_features,labels['label'],train_size=0.9,shuffle=True)

log.fit(xtrain,ytrain)
roc_auc_score(ytest,log.predict_proba(xtest)[:,1])

log.fit(stacked_features,labels['label'])

final_pred = log.predict_proba(stacked_features_p)[:,1]

submit = kaggle.copy(deep=True)
submit['label'] = final_pred

submit.set_index('person').to_csv('submit.csv')
#no hubo mejora en el submit

## XGBoost con de todo

import scipy.stats as stats

from sklearn.model_selection import RandomizedSearchCV

params={
    'n_estimators':stats.randint(150,500),
    'learning_rate':stats.uniform(0.01,0.3),
    'subsample':stats.uniform(0.3,0.7),
    'min_child_weight':[1,5,10],
    'max_depth':[3,10,6],
    'gamma':stats.randint(0,10),
    'colsample_bytree':stats.uniform(0.4,0.6)
}
grid = RandomizedSearchCV(xgb.XGBClassifier(n_jobs=-1),
                          param_distributions=params,
                          scoring='roc_auc',
                          cv=2,
                          verbose=1,
                          n_iter=100)

forxgboost2 = stacked_features.reset_index().join(encoded).set_index('person')

xgboost2predict = stacked_features_p.reset_index().join(encoded_p).set_index('person')

xtrain,xtest,ytrain,ytest = train_test_split(forxgboost2,labels['label'])

grid.fit(xtrain,ytrain)
roc_auc_score(ytest,grid.predict_proba(xtest)[:,1])

grid.best_estimator_.fit(forxgboost2,labels['label'])

submit = kaggle.copy(deep=True).set_index('person')
submit['label'] = grid.best_estimator_.predict_proba(xgboost2predict)[:,1]

submit.to_csv('submit.csv')
#no hubo mejora en el submit

## Logistic regression tuneando parametros (score: 0.85527)

logreg = LogisticRegression()
params = {
    'solver':['newton-cg', 'lbfgs', 'liblinear', 'sag', 'saga'],
}
log = RandomizedSearchCV(logreg,param_distributions=params,verbose=0,cv=4,n_iter=5)

stacked_training.to_csv('preentrenados/training_logrec')

stacked_predict.to_csv('preentrenados/predict_logrec')

xtrain,xtest,ytrain,ytest = train_test_split(stacked_training,labels['label'])

log.fit(xtrain,ytrain)
roc_auc_score(ytest,log.predict_proba(xtest)[:,1])

log.best_estimator_.fit(stacked_training,labels['label'])

final_pred = log.best_estimator_.predict_proba(stacked_predict)[:,1]

submit = kaggle.copy(deep=True)
submit['label'] = final_pred

submit.set_index('person').to_csv('submit.csv')
