==> 23_catboost-normalizado.ipynb <==
from sklearn.metrics import roc_auc_score

import numpy as np
import scipy.stats as stats

import xgboost as xgb
import catboost as cb

import pandas as pd
from sklearn.model_selection import train_test_split
import category_encoders

from sklearn.model_selection import RandomizedSearchCV

kaggle = pd.read_csv('trocafone_kaggle_test.csv')

labels = pd.read_csv('labels_training_set.csv')

features_06 = pd.read_csv('features_test.csv')

features_04 = pd.read_csv('features_04.csv')

repeated = ['busco_productos', 'clickeo_ad', 'hizo_checkout', 'entro_desde_motor',
       'hizo_conversion', 'filtro_busquedas', 'busco_marca', 'visito_sitio',
       'hizo_lead','primer_visita','ultima_visita']

processed = features_06.set_index('person').join(features_04.set_index('person').drop(repeated,axis=1),lsuffix='1')

cat_features = []
for i,col in enumerate(processed.columns):
    if processed[col].dtype.name=='object':
        cat_features.append(i)

predict_features = kaggle.set_index('person').join(processed)

training_features = labels.set_index('person').join(processed)
training_labels = training_features['label']
training_features = training_features.drop(['label'],axis=1)

## Catboost normal

xtrain,xtest,ytrain,ytest = train_test_split(training_features,training_labels)

catBoost = cb.CatBoostClassifier(iterations=200,
                                 depth=8,
                                 l2_leaf_reg=50,
                                 random_strength=0.1,
                                 bagging_temperature=0,
                                 border_count=128)

catBoost.fit(xtrain,ytrain,verbose=0,cat_features=cat_features)

catBoost.save_model('86436catboost')

roc_auc_score(ytest,catBoost.predict_proba(xtest)[:,1])

catBoost.fit(training_features,training_labels,verbose=0,cat_features=cat_features)

pred = catBoost.predict_proba(predict_features)[:,1]
submit = kaggle.set_index('person')
submit['label'] = pred

submit.head()

submit.to_csv('submit.csv')

## Random subsets

from sklearn.util,bvs.random import sample_without_replacement

xtrain.index.size

cbClassifier = cb.CatBoostClassifier(iterations=200,
                                 depth=8,
                                 l2_leaf_reg=50,
                                 random_strength=0.1,
                                 bagging_temperature=0,
                                 border_count=128)
def tryRandomSubsets (xtrain,xtest,ytrain,ytest,n):
    results = []
    for i in range(n):
        indices = sample_without_replacement(n_population=14560,n_samples=6000)
        cbClassifier.fit(xtrain.iloc[indices],ytrain.iloc[indices],cat_features=cat_features,verbose=0)
        score = roc_auc_score(ytest,cbClassifier.predict_proba(xtest)[:,1])
        results.append([indices,score])
    return results


res = tryRandomSubsets(xtrain,xtest,ytrain,ytest,10)

res

res[7]

cbClassifier.fit(xtrain.iloc[res[7][0]],ytrain.iloc[res[7][0]],cat_features=cat_features,verbose=0)

submit = kaggle.set_index('person')
submit['label'] = cbClassifier.predict_proba(predict_features)[:,1]

submit.to_csv('submit.csv')

## Dropeando las unknown (score: 0.86514)

from sklearn.preprocessing import StandardScaler

cb_features = training_features.drop(['Unknown','Unknown.1'],axis=1)
cb_predict = predict_features.drop(['Unknown','Unknown.1'],axis=1)                                  

not_numeric = []
numeric = []
for col in cb_features.columns:
    if cb_features[col].dtype.name == 'object':
        not_numeric.append(col)
    else:
        numeric.append(col)

numeric_features_t = cb_features.drop(not_numeric,axis=1)
not_numeric_features_t = cb_features.drop(numeric,axis=1)

numeric_features_p = cb_predict.drop(not_numeric,axis=1)
not_numeric_features_p = cb_predict.drop(numeric,axis=1)

scaler = StandardScaler()
estandarizadas_t = scaler.fit_transform(numeric_features_t,training_labels)
estandarizadas_p = scaler.transform(numeric_features_p)

for_training = not_numeric_features_t.reset_index().join(pd.DataFrame(estandarizadas_t)).set_index('person')

for_predict = not_numeric_features_p.reset_index().join(pd.DataFrame(estandarizadas_p)).set_index('person')

xtrain,xtest,ytrain,ytest = train_test_split(for_training,training_labels,train_size=0.6)

cb_classifier = cb.CatBoostClassifier(iterations=400,
                                 depth=8,
                                 l2_leaf_reg=50,
                                 random_strength=0.1,
                                 bagging_temperature=0,
                                 border_count=128)

cat_features2 = []
for i,col in enumerate(xtrain.columns):
    if xtrain[col].dtype.name=='object':
        cat_features2.append(i)

cb_classifier.fit(xtrain,ytrain,cat_features=cat_features2,verbose=0)

roc_auc_score(ytest,cb_classifier.predict_proba(xtest)[:,1])

cb_classifier.fit(for_training,training_labels,cat_features=cat_features2,verbose=0)

proba_loco=cb_classifier.predict_proba(for_predict)[:,1]

submit = kaggle.set_index('person')
submit['label'] = proba_loco

submit.to_csv('submit.csv')

cb_classifier.save_model('preentrenados/86514catboost(normalizado)')
