==> 26_features.ipynb <==
import pandas as pd

file = pd.read_csv('events_up_to_01062018.csv',low_memory=False)

file.info()

Separar condition y storage segun si vio o compro <br>
Columnas de los modelos mas comprados y si la persona lo vio o no


conditions = file['condition'].dropna().unique()
storages = file['storage'].dropna().unique()
columns = []

for c in conditions:
    columns.append(c+'_compro')
    columns.append(c+'_vio')
for s in storages:
    columns.append(s+'_compro')
    columns.append(s+'_vio')
    
def viocompro(x):
    a = []
    df_c = x.loc[x.event=='conversion',['condition','storage']]
    df_v = x.loc[x.event=='viewed product',['condition','storage']]
    for c in conditions:
        a.append(c in list(df_c['condition']))
        a.append(c in list(df_v['condition']))
    for s in storages:
        a.append(s in list(df_c['storage']))
        a.append(s in list(df_v['storage']))
    return pd.Series(a)
viocompro = file.groupby('person').apply(viocompro)
viocompro.columns = columns


compras = file.loc[file.event=='conversion','model']
mascomprados = list(compras.value_counts().head(7).index)
def violosmascomprados (x):
    a = []
    b = x.loc[x.event=='viewed product','model'].values
    for c in mascomprados:
        a.append(c in b)
    return pd.Series(a)
mascomprados = file.groupby('person').apply(violosmascomprados)

mascomprados.columns = list(compras.value_counts().head(7).index)

mascomprados.describe()

## Exporto el nuevo archivo de features

features_1 = pd.read_csv('features_05.csv')

features_2 = pd.read_csv('features_test.csv')

repeated = ['busco_productos', 'clickeo_ad', 'hizo_checkout', 'entro_desde_motor',
       'hizo_conversion', 'filtro_busquedas', 'busco_marca', 'visito_sitio',
       'hizo_lead','primer_visita','ultima_visita','entro_direct','uso_computer']

features_06 = features_1.set_index('person').drop(repeated,axis=1).join(features_2.set_index('person'),lsuffix='_1')

features_06 = features_new.drop(['Unknown','Unknown.1'],axis=1)

features_07 = features_new.join(mascomprados).join(viocompro).drop(['Excelente', 'Muito Bom', 'Bom', 'Bom - Sem Touch ID', 'Novo',
       '32GB', '64GB', '128GB', '16GB', '8GB', '256GB', '4GB', '512MB'],axis=1)

features_07.to_csv('features_07.csv')

## Agregar cantidad de eventos por usuario

features_7 = pd.read_csv('features_07.csv', low_memory=False)

cols = ['event','condition','storage','staticpage','search_engine','device_type','channel']

colnames =[]
for i in cols:
    values = file[i].dropna().unique()
    colnames.extend(values)

csrcs = list(file[col].dropna().unique() for col in cols)

gpb=file.groupby('person')

def cant_de_evento(x):
    a = []
    for i,v in enumerate(cols):
        for j in csrcs[i]:
            a.append(len(x[x[v] == j]))
    return pd.Series(a)
features = gpb.apply(cant_de_evento)
features.columns = colnames

features.to_csv('features_test2.csv')

features = features.reset_index()

## Exporto el archivo de features sin las columnas de True o False

features_1 = pd.read_csv('features_05.csv')

#features_2 = pd.read_csv('features_test2.csv') Ahora lo tengo en memoria
features_2 = features

repeated = ['busco_productos', 'clickeo_ad', 'hizo_checkout', 'entro_desde_motor',
       'hizo_conversion', 'filtro_busquedas', 'busco_marca', 'visito_sitio',
       'hizo_lead','primer_visita','ultima_visita','entro_direct','uso_computer']

features_06 = features_1.set_index('person').join(features_2.set_index('person'),lsuffix='_1')

features_06 = features_06.drop(['Unknown'],axis=1)

features_08 = features_06.join(mascomprados).join(viocompro)

features_08.to_csv('features_08.csv')

features_08.info()
