==> 29_xgboost.ipynb <==
import numpy as np

import xgboost as xgb

import pandas as pd
from sklearn.model_selection import train_test_split
import category_encoders

features = pd.read_csv("features_11.csv")
kaggle = pd.read_csv('trocafone_kaggle_test.csv')
labels = pd.read_csv('labels_training_set.csv')

training = labels.set_index('person').join(features.set_index('person'))
train_labels = training['label']
train_data = training.drop('label',axis=1)

predict_data = kaggle.set_index('person').join(features.set_index('person'))

encoder = category_encoders.TargetEncoder(smoothing=True)
train_data_enc = encoder.fit_transform(np.array(train_data),np.array(train_labels))
predict_data_enc = encoder.transform(np.array(predict_data))

xg_classifier = XGBClassifier()

xtrain,xtest,ytrain,ytest = train_test_split(train_data_enc,np.array(train_labels))

train = xgb.DMatrix(data=xtrain,label=ytrain)
test = xgb.DMatrix(data=xtest,label=ytest)
data = xgb.DMatrix(data=train_data_enc,label=np.array(train_labels))
pred = xgb.DMatrix(data=predict_data_enc)
params = {'colsample_bytree': 0.9,
 'gamma': 2,
 'learning_rate': 0.016238752458245277,
 'max_depth': 13,
 'min_child_weight': 5,
 'n_estimators': 100,
 'subsample': 0.7847190225361189,
 'eval_metric':'auc'}

bst = xgb.train(params,train,200,[(train,'train'),(test,'eval')])

bst = xgb.train(params,data,200)

xg_prediction = bst.predict(pred)

submit = kaggle.set_index('person')

submit['label'] = xg_prediction

submit.to_csv('submit.csv')

bst.save_model('86605_xgboost_combinar_con_86602_catboost')

## Combinando con catboost (0.86605)

import catboost as cb

cb_classifier = cb.CatBoostClassifier(iterations=400,
                                     depth=8,
                                     l2_leaf_reg=100,
                                     random_strength=0.1,
                                     bagging_temperature=0,
                                     border_count=254,
                                     learning_rate=0.03)

cb_classifier.load_model('preentrenados/86602catboost')

cb_prediction = cb_classifier.predict_proba(predict_data)[:,1]

prediction = 0.9*cb_prediction+0.1*xg_prediction

submit = kaggle.set_index('person')

submit['label'] = prediction

submit.to_csv('submit.csv')
