==> 30_features.ipynb <==
## Limpio el archivo de features de cosas que pueden estar confundiendo al algoritmo

import pandas as pd
import datetime as dt

features = pd.read_csv("features_08.csv",low_memory=False)

features["primer_visita"] = pd.to_datetime(features["primer_visita"])
features["ultima_visita"] = pd.to_datetime(features["ultima_visita"])

ser=features["primer_visita"]
features["primer_visita"]=ser.map(lambda x: x.strftime('%m-%d'))
ser=features["ultima_visita"]
features["ultima_visita"]=ser.map(lambda x: x.strftime('%m-%d'))
features.head(10)

features.to_csv('features_09.csv')

file = pd.read_csv("events_up_to_01062018.csv", low_memory=False)


lista=file.model.unique()
dic={}
cont=0
for elem in lista:
    precio = int(input("{} Precio de {}".format(cont,elem)))
    dic[elem]=dic.get(elem,0)+precio
    cont += 1
    
    

for elem in lista:
    if elem not in dic:
        precio = int(input("{} Precio de {}".format(cont,elem)))
        dic[elem]=dic.get(elem,0)+precio
        cont += 1

a.pop("NaN")
a=pd.DataFrame.from_dict(dic, orient='index')

a.to_csv("precios_celulares.csv")
