==> 32_features.ipynb <==
import pandas as pd

file = pd.read_csv('events_up_to_01062018.csv',low_memory=False)

features_1 = pd.read_csv('features_10.csv')

features_1 = features_1.set_index('person')

precios = pd.read_csv('precios_celulares.csv')

precios.columns=['model','precio']

## Promedio de precios de una region

new_file = file.merge(precios,on='model',how='left')

new_file = new_file.set_index('person').join(features_1,how='left',lsuffix='_left', rsuffix='_right')

a=new_file.loc[new_file['event']=='viewed product'].groupby('region_sin_unknown')["precio"].mean().to_frame().reset_index()
a.columns=['region_sin_unknown','precio_visto_promedio_region']

b=new_file.loc[new_file['event']=='conversion'].groupby('region_sin_unknown')["precio"].mean().to_frame().reset_index()
b.columns=['region_sin_unknown','precio_compras_promedio_region']

features_1=features_1.reset_index()

features_1= features_1.merge(a,on="region_sin_unknown",how='left')
features_1= features_1.merge(b,on="region_sin_unknown",how='left')

## Diferencia de precios entre el promedio de la region y el usuario

features_1["diferencia_con_la_region"] = features_1['precio_visto_promedio_region']-features_1["promedio_precio_celulares_vistos"]

features_1["diferencia_con_la_region_n"] = (features_1["diferencia_con_la_region"] - features_1["diferencia_con_la_region"].mean() ) / features_1["diferencia_con_la_region"].std()

features_1.head()

features_1.to_csv("features_11.csv")

## Clean features que no aportan

features = pd.read_csv("features_11.csv", low_memory=False)

labels = pd.read_csv("labels_training_set.csv", low_memory=False)

labels["label"].value_counts()

#### Listado de features hasta ahora

for i in features.columns:
    print (i)

lista_no_sirven=[]

lista_no_sirven.append("Ask")

features["Unknown.1"].value_counts()

inn=features.loc[features["Unknown.1"].isin([4,2,1])]
inn

No se sabe bien que es es esta columna, ni porque aparece pero la mayoria de sus datos son cero.
Me fijo si realizaron una conversion las personas con los valores no nulos


joined=labels.set_index("person").join(inn.set_index("person"),how="right")

joined

Su aparicion no pareciera tener nada que ver con la columna de compra

lista_no_sirven.append("Unknown.1")

lista_no_sirven.append("diferencia_con_la_region_n")

lista_no_sirven.append("Unnamed: 0")

features["Bing"].value_counts()

Ok. Aca pareciera haber algo mas interesante, pruebo ver contra los labels

inn = features.loc[features["Bing"] != 0 ]

joined=labels.set_index("person").join(inn.set_index("person"),how="right")
a=joined.dropna().loc[joined["label"]== 1.0]

a["Bing"]

No la borraria ya que algunos de los que buscan por bing terminan dando label 1

lista_automatizada=[]

for elem in features.columns:
    inn = features.loc[features[elem] != 0 ]
    joined=labels.set_index("person").join(inn.set_index("person"),how="right")
    a=joined.dropna().loc[joined["label"]== 1.0]
    if len(a)==0:
        lista_automatizada.append(elem)

lista_automatizada

features=features.drop(lista_automatizada,axis=1)

features= features.drop(lista_no_sirven,axis=1)

features mas importantes (ojo el overfitting)

lista_mas_imp=[]
for elem in features.columns:
    inn = features.loc[features[elem] != 0 ]
    joined=labels.set_index("person").join(inn.set_index("person"),how="right")
    a=joined.dropna().loc[joined["label"]== 1.0]
    if len(a)>912:
        lista_mas_imp.append(elem)

f=features.loc[:,lista_mas_imp].set_index("person")

features = features.set_index("person")
features.head()

features.to_csv("features_12.csv")
