==> 34_GridSearchCatBoost.ipynb <==
from sklearn.metrics import roc_auc_score

from sortedcontainers import SortedList
import copy
import collections
import numpy as np
from itertools import product,chain
import pandas
from sklearn.model_selection import KFold
import catboost as cb

''' a class for doing grid search on a set of parameters provided in a dict. 'pdict' should be a dictionary like the following:
pdict = {'depth':[1,2], 'iterations':[250,100,500], 'thread_count':4}

when grid_search is called it will return an iterator that provides samples from the dictionary e.g.
{'depth':1, 'iterations':250, 'thread_count':4}
{'depth':2, 'iterations':250, 'thread_count':4}
{'depth':1, 'iterations':100, 'thread_count':4}
etc.
after calling an iteration of grid_search, you need to test the classifier and run 'register_result'
This will update the internal list of results, so that the next call to grid_search will use the best
parameters for all the parameters not currently being updated.

grid_search can be provided a list e.g. grid_search(['depth']) this will use the current best parameters for all
the other arguments and only search over 'depth'. You can then call e.g. grid_search(['iterations']) and it will use
the best depth found previously and cycle through all the 'iterations'. Searching incrementally can be much faster
than doing a full grid search, but may miss the global optimum. '''
class paramsearch:
    def __init__(self,pdict):    
        self.pdict = {}
        # if something is not passed in as a sequence, make it a sequence with 1 element
        #   don't treat strings as sequences
        for a,b in pdict.items():
            if isinstance(b, collections.Sequence) and not isinstance(b, str): self.pdict[a] = b
            else: self.pdict[a] = [b]
        # our results are a sorted list, so the best score is always the final element
        self.results = SortedList()       
                    
    def grid_search(self,keys=None):
        # do grid search on only the keys listed. If none provided, do all
        if keys==None: keylist = self.pdict.keys()
        else: keylist = keys
 
        listoflists = [] # this will be list of lists of key,value pairs
        for key in keylist: listoflists.append([(key,i) for i in self.pdict[key]])
        for p in product(*listoflists):
            # do any changes to the current best parameter set
            if len(self.results)>0: template = self.results[-1][1]
            else: template = {a:b[0] for a,b in self.pdict.items()}
            # if our updates are the same as current best, don't bother
            if self.equaldict(dict(p),template): continue
            # take the current best and update just the ones to change
            yield self.overwritedict(dict(p),template)
                              
    def equaldict(self,a,b):
        for key in a.keys(): 
            if a[key] != b[key]: return False
        return True            
                              
    def overwritedict(self,new,old):
        old = copy.deepcopy(old)
        for key in new.keys(): old[key] = new[key]
        return old            
    
    # save a (score,params) pair to results. Since 'results' is a sorted list,
    #   the best score is always the final element. A small amount of noise is added
    #   because sorted lists don't like it when two scores are exactly the same    
    def register_result(self,result,params):
        self.results.add((result+np.random.randn()*1e-10,params))    
        
    def bestscore(self):
        return self.results[-1][0]
        
    def bestparam(self):
        return self.results[-1][1]

def crossvaltest(params,train_set,train_label,cat_dims,n_splits=3):
    kf = KFold(n_splits=n_splits,shuffle=True) 
    res = []
    for train_index, test_index in kf.split(train_set):
        train = train_set.iloc[train_index,:]
        test = train_set.iloc[test_index,:]

        labels = train_label.iloc[train_index]
        test_labels = train_label.iloc[test_index]

        clf = cb.CatBoostClassifier(**params)
        clf.fit(train, np.ravel(labels), cat_features=cat_dims)

        res.append(np.mean(clf.predict(test)==np.ravel(test_labels)))
    return np.mean(res)

import numpy as np
import scipy.stats as stats

import xgboost as xgb
import catboost as cb

import pandas as pd
from sklearn.model_selection import train_test_split
import category_encoders

from sklearn.model_selection import KFold
from itertools import product,chain

from sklearn.model_selection import RandomizedSearchCV

#### Levanto los datos de csv

kaggle = pd.read_csv('trocafone_kaggle_test.csv')

labels = pd.read_csv('labels_training_set.csv')

features_1 = pd.read_csv("features_08.csv")

processed = features_1.set_index('person')

# Indico que columnas son categoricas
cat_features = []
for i,col in enumerate(processed.columns):
    if processed[col].dtype.name=='object':
        cat_features.append(i)

#Junto los labels 
predict_features = kaggle.set_index('person').join(processed)

#Datos de entrenamiento
training_features = labels.set_index('person').join(processed)

#Solo las labels de entrenamiento
training_labels = training_features['label']

#Datos de entrenamiento sin labels
training_features = training_features.drop(['label'],axis=1)

#### Definicion de funciones para la grid search

params = {'depth':[3,1,2,6,4,5,7,8,9,10],
          'iterations':[250,100,500,1000],
          'learning_rate':[0.03,0.001,0.01,0.1,0.2,0.3], 
          'l2_leaf_reg':[3,1,5,10,100],
          'border_count':[32,5,10,20,50,100,200],
          'border_count':[50,5,10,20,100,200],
          'thread_count':4}

def catboost_param_tune(params,train_set,train_label,cat_dims=None,n_splits=3):
    ps = paramsearch(params)
    # search 'border_count', 'l2_leaf_reg' etc. individually 
    #   but 'iterations','learning_rate' together
    for prms in chain(ps.grid_search(['border_count']),
                      ps.grid_search(['border_count']),
                      ps.grid_search(['l2_leaf_reg']),
                      ps.grid_search(['iterations','learning_rate']),
                      ps.grid_search(['depth'])):
        res = crossvaltest(prms,train_set,train_label,cat_dims,n_splits)
        # save the crossvalidation result so that future iterations can reuse the best parameters
        ps.register_result(res,prms)
        print(res,prms,'best:',ps.bestscore(),ps.bestparam())
    return ps.bestparam()

#Split de datos de entrenamiento y test
xtrain,xtest,ytrain,ytest = train_test_split(training_features,training_labels)

bestparams = catboost_param_tune(params,training_features,training_labels,cat_features)

bestparams

best={'depth': 3,
 'iterations': 1000,
 'learning_rate': 0.03,
 'l2_leaf_reg': 3,
 'border_count': 20,
 'thread_count': 4}

### Pruebo los parametros

xtrain,xtest,ytrain,ytest = train_test_split(training_features,training_labels)

cb_classifier = cb.CatBoostClassifier(**best)

cb_classifier.fit(xtrain,ytrain,cat_features=cat_features,verbose=0)

cb_classifier.eval_metrics(cb.Pool(xtest,ytest,cat_features=cat_features),'AUC',plot=True)

roc_auc_score(ytest,cb_classifier.predict_proba(xtest)[:,1])

cb_classifier.fit(training_features, np.ravel(training_labels), cat_features=cat_features,verbose=0)


cb_classifier.save_model("preentrenados/cb_f08")

proba_loco=cb_classifier.predict_proba(predict_features)[:,1]

submit = kaggle.set_index('person')
submit['label'] = proba_loco

submit.to_csv('submit.csv')

res = clf.predict(test_set)
print('error:',1-np.mean(res==np.ravel(test_label)))
