==> 39_bagging.ipynb <==
import numpy as np

import sklearn.ensemble as en

import pandas as pd
from sklearn.model_selection import train_test_split
import category_encoders

import scipy.stats as st

from sklearn.model_selection import RandomizedSearchCV

import pandas as pd

kaggle = pd.read_csv('trocafone_kaggle_test.csv')

labels = pd.read_csv('labels_training_set.csv')

processed = pd.read_csv('features_10.csv')

processed = processed.set_index('person').fillna(0)
processed.head()

training_features = labels.set_index('person').join(processed)
predict_features = kaggle.set_index('person').join(processed)
training_labels = training_features['label']
training_features = training_features.drop('label',axis=1)

training_features.head()

## Mean encoding

tg = category_encoders.TargetEncoder()
encoded_training = tg.fit_transform(np.array(training_features),np.array(training_labels))
encoded_predict = tg.transform(np.array(predict_features))

#train_features_norm = StandardScaler().fit_transform(encoded_training)
#predict_features_norm = StandardScaler().fit_transform(encoded_predict)

params={
    'n_estimators':st.randint(10,50),
    'max_samples':st.uniform(0.1,0.9),
    'max_features':st.uniform(0.1,0.9),
    
}
grid = RandomizedSearchCV(en.BaggingClassifier(oob_score=True,bootstrap=True,bootstrap_features=True),param_distributions=params,scoring='roc_auc',cv=2,verbose=2,n_iter=25)

xtr,xte,ytr,yte=train_test_split(encoded_training,training_labels)

grid.fit(xtr,ytr)

grid.score(X=xte,y=yte)
