==> 41_features_limpieza.ipynb <==
import pandas as pd

def ver_cols (a):
    for i,v in enumerate(a.columns):
        print(str(i)+' - '+v+' - '+a[v].dtype.name)

def duplicate_columns(frame):
    groups = frame.columns.to_series().groupby(frame.dtypes).groups
    dups = []

    for t, v in groups.items():

        cs = frame[v].columns
        vs = frame[v]
        lcs = len(cs)

        for i in range(lcs):
            iv = vs.iloc[:,i].tolist()
            for j in range(i+1, lcs):
                jv = vs.iloc[:,j].tolist()
                if iv == jv:
                    dups.append(cs[i])
                    break

    return dups

features_1 = pd.read_csv('features_12.csv')

file = pd.read_csv('events_up_to_01062018.csv',low_memory=False)

file.info()

## Limpieza de features

ver_cols(features_1)

features_2 = features_1.drop(features_1.columns[59:84],axis=1)

## Limpieza de duplicados

features_3 = features_2.drop(duplicate_columns(features_2),axis=1)

## Encoding de marca del celular segun precio promedio

features_3['marca_celular_mas_visto'] = features_2['prod_visto'].str.split(' ').str[0]

promedio_marcas = features_3.loc[:,['marca_celular_mas_visto','precio_del_cel_mas_visto']].groupby('marca_celular_mas_visto').mean()

promedio_marcas.columns=['precio_prom_de_marca_de_cel_mas_visto']

features_4 = features_3.merge(promedio_marcas.reset_index(),on='marca_celular_mas_visto',how='left')

features_4

## Dropeo prod_visto porque ya esta encodeada

features_5 = features_4.drop('prod_visto',axis=1)

ver_cols(features_5)

## Renombro las columnas 0,1,2,3,... por lo que son

features_6 = features_5.rename({'0':'vio_galaxy_J5','1':'vio_iphone5s','2':'vio_iphone6','3':'vio_iphone6s','4':'vio_motog2_3gdual','5':'vio_galaxy_J7Prime','6':'vio_motog4plus'},axis=1)

ver_cols(features_6)

features_test = pd.read_csv('features_test.csv').drop(['Unknown','Unknown.1','Excelente',
       'Muito Bom', 'Bom', 'Bom - Sem Touch ID', 'Novo', '32GB', '64GB',
       '128GB', '16GB', '8GB', '256GB', '4GB', '512MB', 'Conditions',
       'AboutUs', 'how-to-sell', 'trust-trocafone', 'how-to-buy',
       'CustomerService', 'club-trocafone', 'Quiosks', 'FaqEcommerce',
       'TermsAndConditionsReturnEcommerce', 'TermsAndConditionsEcommerce',
       'galaxy-s8', 'PrivacyEcommerce', 'black_friday'],axis=1)

features_7 = features_6.set_index('person').join(features_test.set_index('person'),how='left',rsuffix='_bool')

features_8 = features_7.drop(duplicate_columns(features_7),axis=1)

## Feature temporal en segundos

(file['timestamp_b'].max() - ene12018).seconds

ene12018 = pd.to_datetime('01-01-2018')
file['timestamp_b'] = pd.to_datetime(file.timestamp)
def minutos_desde_inicio_2018 (x):
    primer_visita = (x['timestamp_b'].min() - ene12018).seconds
    ultima_visita = (x['timestamp_b'].max() - ene12018).seconds
    return pd.Series([primer_visita,ultima_visita])
pr_ult_visita = file.groupby('person').apply(minutos_desde_inicio_2018)


pr_ult_visita.columns = ['primer_visita_seconds','ultima_visita_seconds']

features_9 = features_8.join(pr_ult_visita,how='left')

features_9.to_csv('features_13.csv')

ver_cols(features_9)
