==> 42-xgboost.ipynb <==
import xgboost as xgb
import pandas as pd
import numpy as np
from sklearn.metrics import roc_auc_score
from scipy.stats import uniform,randint
from sklearn.model_selection import RandomizedSearchCV
from sklearn.model_selection import StratifiedKFold

from sklearn.model_selection import train_test_split
import category_encoders

features = pd.read_csv('features_14_reducido.csv')

features.info()

kaggle = pd.read_csv('trocafone_kaggle_test.csv')
labels = pd.read_csv('labels_training_set.csv')

training = labels.set_index('person').join(features.set_index('person'))
train_labels = training['label']
train_data = training.drop('label',axis=1)

predict_data = kaggle.set_index('person').join(features.set_index('person'))

encoder = category_encoders.TargetEncoder()
train_data_enc = encoder.fit_transform(np.array(train_data),np.array(train_labels))
predict_data_enc = encoder.transform(np.array(predict_data))

## Random Search

xtrain,xtest,ytrain,ytest = train_test_split(train_data_enc,np.array(train_labels),train_size=0.75,stratify=np.array(train_labels))

params={
    'n_estimators':randint(150,500),
    'learning_rate':uniform(0.01,0.3),
    'subsample':uniform(0.3,0.7),
    'min_child_weight':randint(2,20),
    'max_depth':randint(3,15),
    'gamma':randint(0,10),
    'colsample_bytree':uniform(0.4,0.6),
    'subsample':uniform(0,1),
    'max_delta_step':randint(0,5),
    'scale_pos_weight':uniform(0.1,2),
    
}
grid = RandomizedSearchCV(xgb.XGBClassifier(n_jobs=-1,booster='dart'),
                          param_distributions=params,
                          scoring='roc_auc',
                          cv=2,
                          verbose=1,
                          n_iter=25)

grid.fit(xtrain,ytrain)

grid.score(xtest,ytest)

grid.best_params_

## Pruebo con otro split

xg_classifier = grid.best_estimator_

xtr,xte,ytr,yte = train_test_split(train_data_enc,np.array(train_labels),train_size=0.5,stratify=np.array(train_labels))

xg_classifier.fit(xtr,ytr)

roc_auc_score(ytr,xg_classifier.predict_proba(xtr)[:,1])

roc_auc_score(yte,xg_classifier.predict_proba(xte)[:,1])

## Hay overfitting, necesito reducirlo

params_ = {'colsample_bytree': 0.6,
 'gamma': 7,
 'learning_rate': 0.02,
 'max_delta_step': 1,
 'max_depth': 3,
 'min_child_weight': 6,
 'n_estimators': 246,
 'scale_pos_weight': 1.9522454207022248,
 'subsample': 0.4}
xgb_classifier = xgb.XGBClassifier(**params_)

skf = StratifiedKFold(n_splits=10,shuffle=True)

for tr,te in skf.split(train_data_enc,np.array(train_labels)):
    xtr = train_data_enc.iloc[tr,:]
    xte = train_data_enc.iloc[te,:]
    ytr = train_labels[tr]
    yte = train_labels[te]
    xgb_classifier.fit(xtr,ytr)
    print ('Train score: '+ str(roc_auc_score(ytr,xgb_classifier.predict_proba(xtr)[:,1])))
    print ('Test score: '+ str(roc_auc_score(yte,xgb_classifier.predict_proba(xte)[:,1])))
    print (' ')

xgb_classifier.fit(train_data_enc,train_labels)

roc_auc_score(train_labels,xgb_classifier.predict_proba(train_data_enc)[:,1])

## Submit

xgb_classifier.fit(train_data_enc,train_labels)

proba_loco = xgb_classifier.predict_proba(predict_data_enc)[:,1]

submit = kaggle.set_index('person')

submit['label'] = proba_loco

submit.to_csv('submitXG.csv')

proba_loco.mean()
