==> 43-stacking.ipynb <==
import pandas as pd
import category_encoders as ce
import numpy as np

from mlxtend.classifier import StackingCVClassifier
from sklearn.ensemble import RandomForestClassifier
from sklearn.neighbors import KNeighborsClassifier
import catboost as cb
import xgboost as xgb
from sklearn.linear_model import LogisticRegression
from sklearn.pipeline import Pipeline
from sklearn.preprocessing import StandardScaler

features = pd.read_csv('features_13.csv').set_index('person')

kaggle = pd.read_csv('trocafone_kaggle_test.csv')
labels = pd.read_csv('labels_training_set.csv')

training = labels.set_index('person').join(features)
train_labels = training['label']
train_data = training.drop('label',axis=1)

predict_data = kaggle.set_index('person').join(features)

## Stacking

cfeats = []
for i,v in enumerate(features.columns):
    if features[v].dtype.name == 'object':
        cfeats.append(i)

cb_classifier = cb.CatBoostClassifier(iterations=400,
                                 depth=8,
                                 l2_leaf_reg=50,
                                 random_strength=0.1,
                                 bagging_temperature=0,
                                 border_count=128,
                                 cat_features=cfeats)

xg_params = {'colsample_bytree': 0.7476502755257962,
 'gamma': 1,
 'learning_rate': 0.015344184565596445,
 'max_delta_step': 0,
 'max_depth': 11,
 'min_child_weight': 11,
 'n_estimators': 324,
 'subsample': 0.7718583635196341}
xg_classifier = xgb.XGBClassifier(**xg_params)

knn_classifier = KNeighborsClassifier(p=2,n_neighbors=10)

rf_classsifier = RandomForestClassifier(n_estimators=100)

logreg_classifier = LogisticRegression()

p_xgb = Pipeline(steps=[('encoding',ce.TargetEncoder()),('fit',xg_classifier)])

p_knn = Pipeline(steps=[('encoding',ce.TargetEncoder()),('normalization',StandardScaler()),('fit',knn_classifier)])

p_rf = Pipeline(steps=[('encoding',ce.TargetEncoder()),('fit',rf_classsifier)])

stack = StackingCVClassifier(classifiers=[p_xgb,cb_classifier,p_knn,p_rf],meta_classifier=logreg_classifier,use_probas=True)

stack.fit(np.array(train_data.fillna(0)),np.array(train_labels))

prediction = stack.predict_proba(np.array(predict_data.fillna(0)))[:,1]

prediction.mean()

submit = kaggle.set_index('person')
submit['label'] = prediction
submit.to_csv('submit.csv')
