==> 44_KNN + XGBoost.ipynb <==
import numpy as np

from sklearn.metrics import roc_auc_score

import sklearn.neighbors as nb
from sklearn.preprocessing import StandardScaler

import pandas as pd
from sklearn.model_selection import train_test_split
import category_encoders

from sklearn.decomposition import PCA

from sklearn.model_selection import GridSearchCV

import pandas as pd

kaggle = pd.read_csv('trocafone_kaggle_test.csv')

labels = pd.read_csv('labels_training_set.csv')

processed = pd.read_csv('features_13.csv')

processed = processed.set_index('person').fillna(0)

## Mean Encoding

training_features = labels.set_index('person').join(processed)
predict_features = kaggle.set_index('person').join(processed)
training_labels = training_features['label']
training_features = training_features.drop('label',axis=1)

tg = category_encoders.TargetEncoder()
encoded_training = tg.fit_transform(np.array(training_features),np.array(training_labels))
encoded_predict = tg.transform(np.array(predict_features))

## Aplico una reduccion de dimensiones, de 118 a 90

pca = PCA(n_components = 90)

scaler = StandardScaler()
train_data_pca = pca.fit_transform(scaler.fit_transform(encoded_training.fillna(0)))
predict_data_pca = pca.fit_transform(scaler.fit_transform(encoded_predict.fillna(0)))

## Normalizo

train_features_norm = StandardScaler().fit_transform(train_data_pca)
predict_features_norm = StandardScaler().fit_transform(predict_data_pca)

## Clasificador KNN

clasKNN= nb.KNeighborsClassifier(n_neighbors=10000, weights='uniform', algorithm='auto', leaf_size=30, p=3, metric='minkowski', metric_params=None, n_jobs=None)

xtr,xte,ytr,yte=train_test_split(train_features_norm,training_labels)

clasKNN.fit(xtr,ytr)

roc_auc_score(yte,clasKNN.predict_proba(xte)[:,1])

# GridSearch de parametros de KNN

params={
    'n_neighbors':[50,100,1000,2500,1000,5000],
    'p':[1,2,3,4,5,6,7],
    'metric': ["euclidean","chebyshev","minkowski"]
}
grid = GridSearchCV(nb.KNeighborsClassifier(),params,scoring='roc_auc',cv=2,verbose=0)

xtr,xte,ytr,yte=train_test_split(train_features_norm,training_labels,train_size=0.6,stratify=np.array(training_labels))

grid.fit(xtr,ytr)

grid.score(xte,yte)

grid.best_params_

knn_clasf=grid.best_estimator_

roc_auc_score(yte,knn_clasf.predict_proba(xte)[:,1])

## Pruebo otro split

xtr,xte,ytr,yte=train_test_split(train_features_norm,training_labels,train_size=0.6,stratify=np.array(training_labels))

knn_clasf.fit(train_features_norm,training_labels)

pred = knn_clasf.predict_proba(predict_features_norm)[:,1]

pred

submit = kaggle.set_index('person')

submit['label'] = pred

submitXG = pd.read_csv("submitXG.csv").set_index("person")

submit.columns=["labelKNN"]

submit["labelXG"] = submitXG["label"]

submit["label"] = submit["labelKNN"]*0.3 + submit["labelXG"]*0.7

submit = submit.drop(["labelKNN","labelXG"], axis=1)

submit.to_csv("submit.csv")
