==> 45_randomForest.ipynb <==
import numpy as np

from sklearn.ensemble import RandomForestRegressor

from sklearn.metrics import roc_auc_score

import pandas as pd
from sklearn.model_selection import train_test_split
import category_encoders

from sklearn.model_selection import GridSearchCV

features = pd.read_csv('features_14_reducido.csv').fillna("Ninguno")

kaggle = pd.read_csv('trocafone_kaggle_test.csv')

labels = pd.read_csv('labels_training_set.csv')

features.columns[features.isna().any()].tolist()

training = labels.set_index('person').join(features.set_index('person'))
train_labels = training['label']
train_data = training.drop('label',axis=1)

#Features de las personas de kaggle que tengo que predecir
predict_data = kaggle.set_index('person').join(features.set_index('person'))

encoder = category_encoders.TargetEncoder(smoothing=True)
train_data_enc = encoder.fit_transform(np.array(train_data),np.array(train_labels))
predict_data_enc = encoder.transform(np.array(predict_data))

rf_classifier = RandomForestRegressor(n_estimators=100,min_samples_split=4,min_samples_leaf=7)

xtr,xte,ytr,yte=train_test_split(train_data_enc,train_labels,train_size=0.6,stratify=np.array(train_labels))

rf_classifier.fit(xtr,ytr)

roc_auc_score(yte,rf_classifier.predict(xte))

roc_auc_score(ytr,rf_classifier.predict(xtr))

rf_prediction = rf_classifier.predict(predict_data_enc)

submit = kaggle.set_index('person')

submit['label'] = rf_prediction
