==> 46_features_reduccion.ipynb <==
import pandas as pd

def ver_cols (a):
    for i,v in enumerate(a.columns):
        print(str(i)+' - '+v+' - '+a[v].dtype.name)

def duplicate_columns(frame):
    groups = frame.columns.to_series().groupby(frame.dtypes).groups
    dups = []

    for t, v in groups.items():

        cs = frame[v].columns
        vs = frame[v]
        lcs = len(cs)

        for i in range(lcs):
            iv = vs.iloc[:,i].tolist()
            for j in range(i+1, lcs):
                jv = vs.iloc[:,j].tolist()
                if iv == jv:
                    dups.append(cs[i])
                    break

    return dups

features_1 = pd.read_csv('features_13.csv')

def ver_cols (a):
    for i,v in enumerate(a.columns):
        print(str(i)+' - '+v+' - '+a[v].dtype.name)

def duplicate_columns(frame):
    groups = frame.columns.to_series().groupby(frame.dtypes).groups
    dups = []

    for t, v in groups.items():

        cs = frame[v].columns
        vs = frame[v]
        lcs = len(cs)

        for i in range(lcs):
            iv = vs.iloc[:,i].tolist()
            for j in range(i+1, lcs):
                jv = vs.iloc[:,j].tolist()
                if iv == jv:
                    dups.append(cs[i])
                    break

    return dups

features_1 = pd.read_csv('features_12.csv')

file = pd.read_csv('events_up_to_01062018.csv',low_memory=False)

file = pd.read_csv('events_up_to_01062018.csv',low_memory=False)

ver_cols(features_1)

## Posibles ruidos

a="""12 - opsis - object
13 - scrnres - object
18 - primer_channel - object
19 - personas_mismo_opsis - int64
20 - personas_mismo_scrnres - int64
24 - entro_direct - bool
25 - uso_computer - bool
38 - Google - int64
39 - Yahoo - int64
40 - Bing - int64
41 - Smartphone - int64
42 - Computer - int64
43 - Tablet - int64
44 - Paid - int64
45 - Organic - int64
46 - Direct - int64
47 - Referral - int64
48 - Social - int64
49 - Email - int64
58 - Excelente_vio - bool
59 - Muito Bom_compro - bool
60 - Muito Bom_vio - bool
61 - Bom_compro - bool
62 - Bom_vio - bool
63 - Bom - Sem Touch ID_compro - bool
64 - Bom - Sem Touch ID_vio - bool
65 - Novo_vio - bool
66 - 32GB_compro - bool
67 - 32GB_vio - bool
68 - 64GB_compro - bool
69 - 64GB_vio - bool
70 - 128GB_compro - bool
71 - 128GB_vio - bool
72 - 16GB_compro - bool
73 - 16GB_vio - bool
74 - 8GB_compro - bool
75 - 8GB_vio - bool
76 - 256GB_vio - bool
77 - 4GB_compro - bool
78 - 4GB_vio - bool
79 - 512MB_compro - bool
80 - 512MB_vio - bool
93 - viewed product_bool - bool
94 - searched products_bool - bool
95 - ad campaign hit_bool - bool
96 - staticpage_bool - bool
97 - checkout_bool - bool
98 - search engine hit_bool - bool
99 - conversion_bool - bool
100 - generic listing_bool - bool
101 - brand listing_bool - bool
102 - visited site_bool - bool
103 - lead_bool - bool
104 - Google_bool - bool
105 - Yahoo_bool - bool
106 - Bing_bool - bool
107 - Ask - bool
108 - Smartphone_bool - bool
109 - Computer_bool - bool
110 - Tablet_bool - bool
111 - Paid_bool - bool
112 - Organic_bool - bool
113 - Direct_bool - bool
114 - Referral_bool - bool
115 - Social_bool - bool
116 - Email_bool - bool"""

a=a.split("\n")

dropear=[]
for e in a:
    e=e.split(" - ")
    dropear.append(e[1])


print(dropear)



features_1=features_1.drop(dropear,axis=1)

features_1=features_1.drop("Bom - Sem Touch ID_compro",axis=1)

features_1=features_1.drop("Bom - Sem Touch ID_vio",axis=1)

features_1

features_1.set_index("person").to_csv("features_14_reducido.csv")
